tools/drawer-ocr/pyproject.toml
[project]
name = "drawer-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "pillow-heif>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
